+++ src/flatvectors/FlatPaints.h
@@ -9,16 +9,16 @@
 
 class FlatPaints {
 public:
-    static void setColorPaint(fvUniform& uniform, long color);
-    static void setImagePaint(fvUniform& uniform, float* affineImg, long color, int cycleMethod);
+    static void setColorPaint(fvUniform& uniform, int32 color);
+    static void setImagePaint(fvUniform& uniform, float* affineImg, int32 color, int32 cycleMethod);
     static void setLinearGradientPaint(fvUniform& uniform, float* affine,
                                        float x1, float y1, float x2, float y2,
-                                       int count, float* stops, long* colors, int cycleMethod);
+                                       int32 count, float* stops, int32* colors, int32 cycleMethod);
     static void setRadialGradientPaint(fvUniform& uniform, float* affine,
                                        float x, float y, float rIn, float rOut, float fx, float fy,
-                                       int count, float* stops, long* colors, int cycleMethod);
+                                       int32 count, float* stops, int32* colors, int32 cycleMethod);
     static void setBoxGradientPaint(fvUniform& uniform, float* affine,
-                                    float x, float y, float w, float h, float r, float f, float a, long c);
+                                    float x, float y, float w, float h, float r, float f, float a, int32 c);
 };
